LayoutTests/fast/forms/old-names-expected.txt

This tests accessing form elements by name. IE only lets you look up names under the first name the element had and does not respond to name changes. Firefox will let each element be looked up under its original name as long as no actual element has that name. This test has been written to expect the Firefox behavior. This original name quirk applies only to the syntax form.name and not to the form.elements collection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS form.length is 2
PASS form.original is a
PASS form.originalB is b
PASS form.second is undefined
PASS form.third is undefined
PASS form.elements.original is a
PASS form.elements.originalB is b
PASS form.elements.second is undefined
PASS form.elements.third is undefined

now change the form item a's name to second

PASS form.length is 2
PASS form.original is a
PASS form.originalB is b
PASS form.second is a
PASS form.third is undefined
PASS form.elements.original is undefined
PASS form.elements.originalB is b
PASS form.elements.second is a
PASS form.elements.third is undefined

now change the form item a's name to third

PASS form.length is 2
PASS form.original is a
PASS form.originalB is b
PASS form.second is a
PASS form.third is a
PASS form.elements.original is undefined
PASS form.elements.originalB is b
PASS form.elements.second is undefined
PASS form.elements.third is a

now change form item b's name to second

PASS form.length is 2
PASS form.original is a
PASS form.originalB is b
PASS form.second is b
PASS form.elements.original is undefined
PASS form.elements.originalB is undefined
PASS form.elements.second is b

now change a form item b's name to third

PASS form.length is 2
PASS form.original is a
PASS form.originalB is b
PASS form.second is b
PASS form.third.length is 2
PASS form.third.item(0) is a
PASS form.third.item(1) is b
PASS form.elements.original is undefined
PASS form.elements.originalB is undefined
PASS form.elements.second is undefined
PASS form.elements.third.length is 2
PASS form.elements.third.item(0) is a
PASS form.elements.third.item(1) is b

now remove element a

PASS form.length is 1
PASS form.original is a
PASS form.originalB is b
PASS form.second is b
PASS form.third is b
PASS form.elements.original is undefined
PASS form.elements.originalB is undefined
PASS form.elements.second is undefined
PASS form.elements.third is b
